% Generated by roxygen2 (4.1.0.9001): do not edit by hand
\name{charmap}
\alias{charmap}
\title{Substituting Strings Against a Mapping Table}
\usage{
charmap(x, maps, nomatch = NULL, partial = FALSE, rev = FALSE)
}
\arguments{
  \item{x}{character vector to match}

  \item{maps}{mapping tables.
May be a character vector, a list of character vectors or a function.}

  \item{nomatch}{character string to be used for non-matched elements of \code{x}.
If \code{NULL}, these elements are left unchanged.}

  \item{partial}{logical that indicates if partial matches are allowed,
in which case mappings are used as regular expressions.}
}
\description{
  Match the elements of a character vectors against a mapping table,
that can specify multiple exact or partial matches.
}

