% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devutils.R
\name{compile_src}
\alias{compile_src}
\title{Compile Source Files from a Development Package}
\usage{
compile_src(pkg = NULL, load = TRUE)
}
\arguments{
\item{pkg}{the name of the package to compile}

\item{load}{a logical indicating whether the compiled library should be loaded
after the compilation (default) or not.}
}
\value{
Returns nothing. Used for its side effect of compiling source files.
}
\description{
Compile Source Files from a Development Package
}
