% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\docType{data}
\name{CRAN}
\alias{CRAN}
\title{Main CRAN Mirror URL}
\format{
An object of class \code{character} of length 1.
}
\usage{
CRAN
}
\description{
\code{CRAN} simply contains the url of CRAN main mirror
(\url{https://cran.r-project.org}), and aims at simplifying its use, e.g., in
calls to \code{\link{install.packages}}.
}
\examples{
\donttest{
install.packages('pkgmaker', repos=CRAN)
}
}
\keyword{datasets}
