% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{digest_function}
\alias{digest_function}
\title{Compute Function Digest Hash}
\usage{
digest_function(fun, n = Inf)
}
\arguments{
\item{fun}{a function}

\item{n}{a single numeric that indicates the length of the hash.}
}
\value{
a character string
}
\description{
Computes a digest hash of the body and signature of a function.
Note that the hash is not affected by attributes or the
function's environment.
}
\details{
The hash itself is computed using \link[digest:digest]{digest::digest}.
}
