\name{.packageMetaRegistry}
\alias{.packageMetaRegistry}
\title{Creates or Retrieves a Package Meta Registry}
\usage{
  .packageMetaRegistry(package, quiet = FALSE,
    create = FALSE)
}
\description{
  This function is used in \code{\link{setPackageRegistry}}
  and \code{\link{packageRegistry}} to create or query meta
  registries.
}
\keyword{internal}

