% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{factor2character}
\alias{factor2character}
\title{Converting Factors to Character Vectors}
\usage{
factor2character(x)
}
\arguments{
\item{x}{\code{data.frame} or \code{ExpressionSet} object}
}
\value{
an object of the same class as \code{x}.
}
\description{
Converts all \code{factor} variables to character vectors in a \code{data.frame}
or phenotypic data.
}
