% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{find_devpackage}
\alias{find_devpackage}
\title{Find Path to Development Package Root Directory}
\usage{
find_devpackage(x, error = TRUE)
}
\arguments{
\item{x}{name of the development package to lookup.}

\item{error}{logical that indicates if an error is thrown when the project root directory
could not be found.}
}
\value{
A character string containing the path to the package.
}
\description{
Development packages are looked-up according to rules
defined in a file \code{.Rpackages} in the user's home directory.
}
\section{Specification of package path}{

Package paths are specified in a list with:
\itemize{
\item unnamed elements: character strings give path to directories to lookup for sub-directories that match
exactly the package's name;
\item named element containing character strings: these are paths that are looked up only for packages that
match the element name. If the element name contains any of the characters \verb{*?()$^\\\][}, then it is matched using
regular expression.
}
}

