% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gfile}
\alias{gfile}
\title{Open a File Graphic Device}
\usage{
gfile(filename, width, height, ...)
}
\arguments{
\item{filename}{path to the image file to create.}

\item{width}{output width}

\item{height}{output height}

\item{...}{other arguments passed to the relevant device function
such as \code{\link{png}} or \code{\link{pdf}}.

importFrom grDevices bmp jpeg pdf png svg tiff}
}
\description{
Opens a graphic device depending on the file extension.
}
