% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hasEnvar}
\alias{hasEnvar}
\title{Check Environment Variables}
\usage{
hasEnvar(x)
}
\arguments{
\item{x}{environment variable name, as a character vector.}
}
\description{
Tells if some environment variable(s) are defined.
}
\examples{
\dontshow{
options(R_CHECK_RUNNING_EXAMPLES_=TRUE) ## roxygen generated flag
}


hasEnvar('_R_CHECK_TIMINGS_')
hasEnvar('ABCD')

}
