% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette.R
\name{inSweave}
\alias{inSweave}
\title{Identifying Sweave Run}
\usage{
inSweave()
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Tells if the current code is being executed within a Sweave document.
}
\examples{
\dontshow{
options(R_CHECK_RUNNING_EXAMPLES_=TRUE) ## roxygen generated flag
}

## inSweave

# Not in a Sweave document 
inSweave()

# Within a Sweave document


}
