% Generated by roxygen2 (4.1.0): do not edit by hand
\name{list.libs}
\alias{libname}
\alias{list.libs}
\title{Library Files Utilities}
\usage{
list.libs(dir, ..., all.platforms = FALSE)

libname(x)
}
\arguments{
  \item{dir}{directory}

  \item{...}{extra arguments passed to \code{\link{list.files}}.}

  \item{all.platforms}{a logical that indicates whether to list library files for
the current platform only (default) or all platforms (Unix, Windows, Mac).}

  \item{x}{a filename}
}
\value{
  a character vector
}
\description{
  Lists binary library files in a directory

\code{libname} extracts library names from a path, removing the
directory part of the path, as well as the platform
specific library extension.
}
\examples{
\dontshow{# roxygen generated flag
options(R_CHECK_RUNNING_EXAMPLES_=TRUE)
}

libname('mylib.so')
libname('/some/path/somewhere/mylib.dll')
}

