\name{listPackageOptions}
\alias{listPackageOptions}
\title{Returns the names of all option sets currently defined.}
\usage{
  listPackageOptions()
}
\value{
  a character vector (possibly empty).
}
\description{
  Returns the names of all option sets currently defined.
}
\examples{
packageOptionSets()
}

