% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette.R
\name{make_vignette_auxfiles}
\alias{make_vignette_auxfiles}
\title{Generate RMarkdown Vignette Axiliary Files}
\usage{
make_vignette_auxfiles(
  PACKAGE,
  input = NULL,
  bibfile = "library.bib",
  Rpkg.prefix = "Rpackage_",
  ...
)
}
\arguments{
\item{PACKAGE}{package name}

\item{input}{vignette source file.
If \code{NULL} then the current file is obtained via a call to \link[knitr:current_input]{knitr::current_input}.}

\item{bibfile}{output file for R package citations.}

\item{Rpkg.prefix}{prefix to use when generating the bibtex entries of cited R packages.
If \code{Rpkg.prefix = 'Rpackage_'}, then Rmardown citations should be \verb{@Rpackage_mypkg}.}

\item{...}{other arguments passed to \link{latex_preamble}}
}
\description{
Generate RMarkdown Vignette Axiliary Files
}
\details{
To use this feature add the following in your YAML header:

\if{html}{\out{<div class="sourceCode">}}\preformatted{header-includes:
  - \\input\{"`r pkgmaker::make_vignette_auxfiles('pkgmaker')`"\}
bibliography: library.bib
}\if{html}{\out{</div>}}
}
