% Generated by roxygen2 (4.1.0): do not edit by hand
\name{makeFakeVignette}
\alias{makeFakeVignette}
\title{Generate a Fake Vignette}
\usage{
makeFakeVignette(src, out, PACKAGE = NULL)
}
\arguments{
  \item{src}{original Sweave file}

  \item{out}{output file}

  \item{PACKAGE}{package name where to look the source vignette}
}
\description{
  Generate a Fake Vignette
}

