\name{packageDependencies}
\alias{packageDependencies}
\title{List Package Dependencies}
\usage{
  packageDependencies(x, all = TRUE, as.list = FALSE)
}
\arguments{
  \item{x}{path to package source directory or file.}

  \item{all}{logical that indicates if all dependencies
  should be returned, or only the required ones.}

  \item{as.list}{logical that indicates if the result
  should be a list with one element per type of
  dependency.}
}
\description{
  List Package Dependencies
}

