\name{packageOptionSets}
\alias{packageOptionSets}
\title{Returns the names of all option sets currently defined.}
\usage{
  packageOptionSets()
}
\value{
  a character vector (possibly empty).
}
\description{
  Returns the names of all option sets currently defined.
}
\examples{
packageOptionSets()
}

