% Generated by roxygen2 (4.1.0.9001): do not edit by hand
\name{packageTestEnv}
\alias{packageTestEnv}
\title{Returns the package internal environment where unit tests are stored.}
\usage{
packageTestEnv(pkg)
}
\arguments{
  \item{pkg}{package name.
If missing the caller's package is assumed.}
}
\description{
  Returns the package internal environment where unit tests are stored.
}

