% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibtex.R, R/package.R
\name{write.bib}
\alias{write.bib}
\alias{pkgmaker-defunct}
\title{Defunct Functions in pkgmaker}
\usage{
write.bib(...)
}
\arguments{
\item{...}{extra arguments}
}
\description{
These functions have been defunct or superseded by other 
functions.
}
