% Generated by roxygen2 (4.1.0.9001): do not edit by hand
\name{read.Rprofile}
\alias{read.Rprofile}
\title{Reads YAML Options Embbeded into a File}
\usage{
read.Rprofile(section, file = "~/.Rprofile", text = NULL)
}
\arguments{
  \item{section}{section name to lookup in the file.
In the file this defined by paired tags \code{"#<section_name>", "#</section_name>"},
or a single tag \code{"#<section_name@file_path>"} that redirect to a YAML file.}

  \item{file}{path to the file to parse. Default is to parse the user's \emph{.Rprofile}.}

  \item{text}{text to parse.
If provided, then argument \code{file} is not used.}
}
\description{
  Reads YAML Options Embbeded into a File
}

