% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{read.yaml_section}
\alias{read.yaml_section}
\title{Reads YAML Options Embbeded into a File}
\usage{
read.yaml_section(section, file = "~/.Rprofile", text = NULL)
}
\arguments{
\item{section}{section name to lookup in the file.
In the file this defined by paired tags \code{"#<section_name>", "#</section_name>"},
or a single tag \code{"#<section_name@file_path>"} that redirect to a YAML file.}

\item{file}{path to the file to parse. Default is to parse the user's \emph{.Rprofile}.}

\item{text}{text to parse.
If provided, then argument \code{file} is not used.}
}
\value{
Returns a list representation of the YAML header
}
\description{
Reads YAML Options Embbeded into a File
}
