\name{setPackageRegistry}
\alias{setPackageRegistry}
\title{Adds a registry to the package}
\usage{
  setPackageRegistry(name, regobj, ..., overwrite = TRUE)
}
\arguments{
  \item{regobj}{a \code{\link[=regobj]{registry}} object or
  a single character string that indicates the class of the
  objects that are stored in the sub-registry.}

  \item{name}{Name of a sub-registry}
}
\description{
  Adds a registry to the package
}

