% Generated by roxygen2 (4.1.0.9001): do not edit by hand
\name{render_notes}
\alias{render_notes}
\title{Renders rmarkdown Documents Using System-Wide Defaults}
\usage{
render_notes(input, output_format = NULL, ..., .config = NULL)
}
\arguments{
  \item{.config}{location of the default options (a YAML file)}
}
\description{
  Renders rmarkdown Documents Using System-Wide Defaults
}

