\name{require.quiet}
\alias{qlibrary}
\alias{require.quiet}
\title{Silently Loading Packages}
\usage{
  require.quiet(...)

  qlibrary(...)
}
\arguments{
  \item{...}{extra arguments passed to
  \code{\link{library}} or \code{\link{require}}.}
}
\description{
  \code{require.quiet} silently requires a package.

  \code{qlibrary} silently loads a package.
}

