\name{requirePackage}
\alias{requirePackage}
\title{Require a Package}
\usage{
  requirePackage(pkg, ...)
}
\arguments{
  \item{pkg}{package name as a character string}

  \item{...}{extra arguments concatenated to for the header
  of the error message}
}
\description{
  Require a package with a custom error message
}

