% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\name{bib_roclet}
\alias{bib_roclet}
\alias{roclet_tags.roclet_bib}
\alias{roclet_process.roclet_bib}
\alias{roclet_clean.roclet_bib}
\title{Bibliography roclet}
\usage{
bib_roclet()

\method{roclet_tags}{roclet_bib}(x)

\method{roclet_process}{roclet_bib}(x, parsed, base_path,
  global_options = list())

\method{roclet_clean}{roclet_bib}(x, base_path)
}
\description{
Bibliography roclet
}
\section{Roxygen tags and commands}{

\itemize{
\item @cite:
\item @bibliography
\item \cite:
}
}

