% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette.R
\name{runVignette}
\alias{runVignette}
\title{Compile a Vignette Object}
\usage{
runVignette(x, ...)
}
\arguments{
\item{x}{an object that represents a vignette}

\item{...}{other arguments passed down to the relevant method.}
}
\value{
Returns the value returned by the vignette compiler.
}
\description{
Compile a Vignette Object
}
