% Generated by roxygen2 (4.1.0.9001): do not edit by hand
\name{Rversion}
\alias{Rversion}
\title{Complete R version}
\usage{
Rversion()
}
\description{
  Returns the complete R version, e.g. 2.15.0
}
\examples{
\dontshow{# roxygen generated flag
options(R_CHECK_RUNNING_EXAMPLES_=TRUE)
}

Rversion()
}

