% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{setClassRegistry}
\alias{setClassRegistry}
\title{Automatic S4 Class for Registry Entries}
\usage{
setClassRegistry(registry, Class, ...)
}
\arguments{
\item{registry}{a registry object}

\item{Class}{name of the class to generate}

\item{...}{extra arguments passed to \code{\link{setClass}}.}
}
\value{
No return value, called to declare a registry class.
}
\description{
Automatic S4 Class for Registry Entries
}
