\name{setPackageRegistry}
\alias{setPackageRegistry}
\title{Adds a registry to the package}
\usage{
  setPackageRegistry(name, regobj, ..., overwrite = TRUE)
}
\arguments{
  \item{name}{Name of the registry}

  \item{regobj}{\code{\link[registry]{registry}} object or
  a character string that indicates the class of single
  objects}
}
\description{
  Adds a registry to the package
}

