\name{checkWarning}
\alias{checkIdenticalRNG}
\alias{checkWarning}
\title{Extra Check Functions for RUnit}
\usage{
  checkWarning(expr, expected = NULL, msg = NULL)

  checkIdenticalRNG(x, y = getRNG(), ...)
}
\arguments{
  \item{expr}{an R expression}

  \item{expected}{expected value or regular expression
  pattern}

  \item{msg}{informative message to add to the error in
  case of failure}

  \item{x,y}{objects from which RNG settings are
  extracted.}

  \item{...}{extra arguments passed to
  \code{\link[rngtools]{rng.equal}} in
  \code{checkIdenticalRNG}.}
}
\description{
  \code{checkWarning} checks if a warning is generated by
  an expression, and optionally follows an expected regular
  expression pattern.

  \code{checkIdenticalRNG} checks if two objects have the
  same RNG settings.
}
\examples{
# check warnings
checkWarning({ warning('ah ah'); 3})
checkWarning({ warning('ah oh ah'); 3}, 'oh')
try( checkWarning(3) )
try( checkWarning({ warning('ah ah'); 3}, 'warn you') )
# check for differences in RNG
set.seed(123)
checkIdenticalRNG(123)
try( checkIdenticalRNG(123, 123) )
try( checkIdenticalRNG(123, 1:3) )
}

