\name{writeUnitVignette}
\alias{writeUnitVignette}
\title{Writes Unit Tests Vignette}
\usage{
  writeUnitVignette(pkg, file, results = NULL,
    check = FALSE)
}
\arguments{
  \item{pkg}{Package name}

  \item{file}{Output Sweave (.Rnw) file}

  \item{results}{result file or output character vector}

  \item{check}{logical that indactes the cal was made from
  R CMD check, in which case the vignette is updated only
  if results of unit tests can be found in the unit test
  output directory, where they would have been generated by
  \code{\link{utest}}.}
}
\description{
  Writes a vignette that contains the results from running
  unit test suites.
}

