% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibtex.R
\name{bibtex}
\alias{bibtex}
\alias{packageReferenceFile}
\alias{package_bibliography}
\title{Bibtex Utilities}
\usage{
packageReferenceFile(PACKAGE = NULL, check = FALSE)

package_bibliography(PACKAGE = NULL, load = FALSE)
}
\arguments{
\item{PACKAGE}{package name. If \code{NULL}, then the name of the calling package is used.}

\item{check}{logical that indicates if the result should be an empty string if the
bibliography file (or package) does not exist.}

\item{load}{logical that specifies if the bibliography should be loaded as
a list of \link[utils:bibentry]{utils::bibentry} objects.}
}
\description{
Utility functions to work with BiBTeX files.
}
\section{Functions}{
\itemize{
\item \code{packageReferenceFile}: returns the path to a package REFERENCES.bib file.

\item \code{package_bibliography}: returns the bibliography associated with a package.
This can
}}

\examples{

packageReferenceFile('pkgmaker')
packageReferenceFile('pkgmaker', check = TRUE)

}
