% Generated by roxygen2 (4.1.0): do not edit by hand
\name{packageReferenceFile}
\alias{packageReferenceFile}
\title{Bibtex Utilities}
\usage{
packageReferenceFile(PACKAGE = NULL)
}
\arguments{
  \item{PACKAGE}{package name}
}
\description{
  \code{packageReferenceFile} returns the path to a package REFERENCES.bib file.
}

