% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibtex.R
\name{packageReferenceFile}
\alias{packageReferenceFile}
\title{Bibtex Utilities}
\usage{
packageReferenceFile(PACKAGE = NULL)
}
\arguments{
\item{PACKAGE}{package name}
}
\description{
\code{packageReferenceFile} returns the path to a package REFERENCES.bib file.
}
