% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cgetAnywhere}
\alias{cgetAnywhere}
\title{Get Anywhere}
\usage{
cgetAnywhere(x)
}
\arguments{
\item{x}{a single character string}
}
\value{
The value of \link{getAnywhere}.
}
\description{
Similar to \code{\link{getAnywhere}}, but looks for the value of its argument.
}
