% Generated by roxygen2 (4.1.0.9001): do not edit by hand
\name{checkPlot}
\alias{checkPlot}
\title{Plot in Unit Tests}
\usage{
checkPlot(expr, msg = NULL, width = 1000, height = NULL)
}
\arguments{
  \item{expr}{expression that generate th eplot}

  \item{msg}{plot msg explaining the plot . It will be used as the caption}

  \item{width}{plot width}

  \item{height}{plot height (not used)}
}
\description{
  Saves a plot in a PNG file that will be included in unit test HTML reports.
}
\keyword{internal}

