\name{compile_src}
\alias{compile_src}
\title{Compile Source Files from a Development Package}
\usage{
  compile_src(pkg = NULL, load = TRUE)
}
\arguments{
  \item{pkg}{the name of the package to compile}

  \item{load}{a logical indicating whether the compiled library should be loaded
after the compilation (default) or not.}
}
\value{
  None
}
\description{
  Compile Source Files from a Development Package
}

