% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{find_devpackage}
\alias{find_devpackage}
\title{Find Path to Development Package Root Directory}
\usage{
find_devpackage(x)
}
\arguments{
\item{x}{name of the development package to lookup.}
}
\description{
Development packages are looked-up according to rules
defined in a file \code{.Rpackages} in the user's home directory.
}
