% Generated by roxygen2 (4.1.0.9001): do not edit by hand
\name{is_package_path}
\alias{is_package_path}
\title{Test for Package Root Directory}
\usage{
is_package_path(x, error = FALSE)
}
\arguments{
  \item{x}{path to the directory to test}

  \item{error}{logical that indicates if an error should be raised
if the directory is not a package directory.}
}
\description{
  Tells if a directory is a package directory, i.e. that it contains
a \code{DESCRIPTION} file.
}

