\name{is_something}
\alias{isFALSE}
\alias{isInteger}
\alias{isNA}
\alias{isNumber}
\alias{is_something}
\alias{isString}
\title{Testing Object Type}
\usage{
  isNA(x)

  isFALSE(x)

  isNumber(x, int.ok = TRUE)

  isInteger(x)

  isString(x)
}
\arguments{
  \item{an}{R object}
}
\description{
  Testing Object Type

  Tests if a variable is exactly NA (logical, character,
  numeric or integer)

  Tests if a variable is exactly FALSE

  Tests if a variable is a single number

  Tests if an object is a single integer

  Test if an object is a single character string
}
\seealso{
  \code{\link{isTRUE}}
}

