% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{iterCount}
\alias{iterCount}
\title{Simple Text Iteration Counter}
\usage{
iterCount(n = 100, i0 = 0, title = "Iterations", extra = NULL,
  verbose = TRUE)
}
\arguments{
\item{n}{number of total steps}

\item{i0}{starting step}

\item{title}{character string to use as title}

\item{extra}{character vector providing extra text to add at each step}

\item{verbose}{logical that toggles the counter}
}
\description{
Simple Text Iteration Counter
}
\examples{
\dontshow{
options(R_CHECK_RUNNING_EXAMPLES_=TRUE) ## roxygen generated flag
}


progress <- iterCount(LETTERS)
sapply(LETTERS, function(x){
 sleep(.1)
	progress()
})
# terminate counter
progress(NULL) 

}
