% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette.R
\name{makeFakeVignette}
\alias{makeFakeVignette}
\title{Generate a Fake Vignette}
\usage{
makeFakeVignette(src, out, PACKAGE = NULL)
}
\arguments{
\item{src}{original Sweave file}

\item{out}{output file}

\item{PACKAGE}{package name where to look the source vignette}
}
\value{
No return value, called to create a vignette file.
}
\description{
Generate a Fake Vignette
}
