\name{makeUnitVignette}
\alias{makeUnitVignette}
\title{Make Vignette for Unit Tests}
\usage{
  makeUnitVignette(pkg,
    file = paste(pkg, "-unitTests.pdf", sep = ""), ...,
    check = FALSE)
}
\arguments{
  \item{pkg}{Package name}

  \item{file}{Output file (.Rnw, .tex, or .pdf)}

  \item{...}{extra arguments passed to \code{\link{utest}}.}

  \item{check}{logical that indactes the cal was made from R CMD check, in which case the vignette
is updated only if results of unit tests can be found in the unit test output directory,
where they would have been generated by \code{\link{utest}}.}
}
\value{
  Result of running unit test suite
}
\description{
  Builds a vignette for unit tests in a package using the
  \code{\link{utest}} and a template vignette file.
}

