% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{messagef}
\alias{messagef}
\title{Generate Formated Message}
\usage{
messagef(fmt, ..., domain = NULL, appendLF = TRUE)
}
\arguments{
\item{fmt}{a character vector of format strings, each of up to 8192 bytes.}

\item{...}{values to be passed into \code{fmt}.  Only logical,
    integer, real and character vectors are supported, but some coercion
    will be done: see the \sQuote{Details} section. Up to 100.}

\item{domain}{see \code{\link{gettext}}.}

\item{appendLF}{logical: should messages given as a character string
    have a newline appended?}
}
\description{
Generate a formatted diagnostic message.
This function is a shortcut for \code{message(sprintf(...))}.
}
\examples{
\dontshow{
options(R_CHECK_RUNNING_EXAMPLES_=TRUE) ## roxygen generated flag
}

## messagef

messagef("Hello \%s number \%i", 'world', 4)


}
\seealso{
\code{\link{sprintf}}, \code{\link{message}}
}
