% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{setBiocMirror}
\alias{setBiocMirror}
\alias{getBiocMirror}
\alias{getBiocRepos}
\alias{setCRANMirror}
\title{Setting Mirrors and Repositories}
\usage{
setBiocMirror(url = "http://www.bioconductor.org", version = NULL,
  unique = TRUE)

getBiocMirror()

getBiocRepos(url = "http://www.bioconductor.org", version = NULL)

setCRANMirror(url = CRAN, unique = TRUE)
}
\arguments{
\item{url}{or Bioconductor mirror url}

\item{version}{version number}

\item{unique}{logical that indicate if duplicated urls or names should be
removed.}
}
\description{
\code{setBiocMirror} sets all Bioconductor repositories (software, data,
annotation, etc.).
so that they are directly available to \code{\link{install.packages}}.
It differs from \code{\link{chooseBioCmirror}} in that it effectively enables
the repositories.
}
\section{Functions}{
\itemize{
\item \code{getBiocMirror}: is a shortcut for \code{getOption('BioC_mirror')}, which
returns the current Bioconductor mirror as used by \code{biocLite}.

\item \code{getBiocRepos}: returns urls to all Bioconductor repositories on a
given mirror.

\item \code{setCRANMirror}: sets the preferred CRAN mirror.
}}

