\name{packageOptions}
\alias{packageOptions}
\title{\code{packageOptions} returns a list of package specific options, and behaves
as the base function \code{\link[base]{options}}.}
\usage{
  packageOptions(..., .DATA = packageName())
}
\description{
  \code{packageOptions} returns a list of package specific
  options, and behaves as the base function
  \code{\link[base]{options}}.
}

