% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitTests.R
\name{packageTestEnv}
\alias{packageTestEnv}
\title{Returns the package internal environment where unit tests are stored.}
\usage{
packageTestEnv(pkg)
}
\arguments{
\item{pkg}{package name.
If missing the caller's package is assumed.}
}
\value{
An environment.
}
\description{
Returns the package internal environment where unit tests are stored.
}
