% Generated by roxygen2 (5.0.1): do not edit by hand
\name{parsePackageCitation}
\alias{parsePackageCitation}
\title{Formatting Package Citations in Sweave/knitr Documents}
\usage{
parsePackageCitation(x)
}
\arguments{
  \item{x}{output document, as a single string.}
}
\description{
  Formatting Package Citations in Sweave/knitr Documents
}

