% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette.R
\name{parsePackageCitation}
\alias{parsePackageCitation}
\title{Formatting Package Citations in Sweave/knitr Documents}
\usage{
parsePackageCitation(x)
}
\arguments{
\item{x}{output document, as a single string.}
}
\description{
Formatting Package Citations in Sweave/knitr Documents
}
