% Generated by roxygen2 (4.1.0.9001): do not edit by hand
\name{write.bib}
\alias{pkgmaker-defunct}
\alias{write.bib}
\title{Defunct Functions in pkgmaker}
\usage{
write.bib(...)
}
\arguments{
  \item{...}{extra arguments}
}
\description{
  These functions have been defunct or superseded by other
functions.
}

