% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R, R/packages.R
\name{pkgmaker-deprecated}
\alias{pkgmaker-deprecated}
\alias{requirePackage}
\title{Deprecated Functions in pkgmaker}
\usage{
requirePackage(pkg, ...)
}
\arguments{
\item{pkg}{package name to load.}

\item{...}{extra arguments}
}
\description{
These functions have been deprecated and will be defunct in the next release.
}
