\name{pkgmaker-deprecated}
\alias{pkgmaker-deprecated}
\alias{requirePackage}
\title{Deprecated Functions in pkgmaker}
\usage{
  requirePackage(pkg, ...)
}
\arguments{
  \item{...}{extra arguments}

  \item{pkg}{package name to load.}
}
\description{
  These functions have been deprecated and will be defunct
  in the next release.
}

