\name{quickinstall}
\alias{quickinstall}
\title{Quick Installation of a Source Package}
\usage{
  quickinstall(path, lib = NULL)
}
\arguments{
  \item{path}{path to the package source directory}

  \item{lib}{installation directory. Defaults to the user's
  R default installation directory.}
}
\description{
  Builds and install a minimal version of a package from
  its source directory.
}

