\name{requireRUnit}
\alias{requireRUnit}
\title{Load RUnit Compatible Package}
\usage{
  requireRUnit(...)
}
\arguments{
  \item{...}{arguments not used.}
}
\value{
  nothing
}
\description{
  Loads the package responsible for the implementation of
  the RUnit framework, choosing amongst \sQuote{RUnitX},
  \sQuote{svUnit} and \sQuote{RUnit}.
}

