% Generated by roxygen2 (4.1.0): do not edit by hand
\name{setClassRegistry}
\alias{setClassRegistry}
\title{Automatic S4 Class for Registry Entries}
\usage{
setClassRegistry(registry, Class, ...)
}
\arguments{
  \item{registry}{a registry object}

  \item{Class}{name of the class to generate}

  \item{...}{extra arguments passed to \code{\link{setClass}}.}
}
\description{
  Automatic S4 Class for Registry Entries
}

