% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{userIs}
\alias{userIs}
\title{Checking R User}
\usage{
userIs(user)
}
\arguments{
\item{user}{the usernames to check for, as a character vector.}
}
\description{
Tests if the current R user is amongst a given set of users.
}
