% Generated by roxygen2 (5.0.1): do not edit by hand
\name{using_something}
\alias{using_collate}
\alias{using_dir}
\alias{using_env}
\alias{using_envvar}
\alias{using_lib}
\alias{using_libpaths}
\alias{using_locale}
\alias{using_options}
\alias{using_par}
\alias{using_path}
\alias{using_something}
\title{Execute code in temporarily altered environment.}
\usage{
using_envvar(new, code, action = "replace")

using_env(new, code)

using_locale(new, code)

using_collate(new, code)

using_dir(new, code)

using_libpaths(new, code)

using_lib(new, code)

using_options(new, code)

using_par(new, code)

using_path(new, code, add = TRUE)
}
\arguments{
  \item{new}{values for setting}

  \item{code}{code to execute in that environment}

  \item{action}{(for \code{using_envvar} only): should new values
\code{"replace"}, \code{"suffix"}, \code{"prefix"} existing environmental
variables with the same name.}

  \item{add}{Combine with existing values? Currently for
\code{\link{using_path}} only. If \code{FALSE} all existing
paths are ovewritten, which don't you usually want.}
}
\description{
  \itemize{
  \item \code{using_dir}: working directory
  \item \code{using_collate}: collation order
  \item \code{using_envvar}: environmental variables
  \item \code{using_libpaths}: library paths, replacing current libpaths
  \item \code{using_lib}: library paths, prepending to current libpaths
  \item \code{using_locale}: any locale setting
  \item \code{using_options}: options
  \item \code{using_path}: PATH environment variable
  \item \code{using_par}: graphics parameters
}
}
\section{Deprecation}{

\code{using_env} will be deprecated in devtools 1.2 and removed in
devtools 1.3
}
\examples{
\dontshow{# roxygen generated flag
options(R_CHECK_RUNNING_EXAMPLES_=TRUE)
}

getwd()
using_dir(tempdir(), getwd())
getwd()

Sys.getenv("HADLEY")
using_envvar(c("HADLEY" = 2), Sys.getenv("HADLEY"))
Sys.getenv("HADLEY")

using_envvar(c("A" = 1),
  using_envvar(c("A" = 2), action = "suffix", Sys.getenv("A"))
)
}

