\name{writeUnitVignette}
\alias{writeUnitVignette}
\title{Writes Unit Tests Vignette}
\usage{
  writeUnitVignette(pkg, file)
}
\arguments{
  \item{pkg}{Package name}

  \item{file}{Output Sweave (.Rnw) file}
}
\description{
  Writes a vignette that contains the results from running
  unit test suites.
}

